Migrate recipe details page to /recipes/[id] route
Key updates:
- Move recipe details page from /[id].vue to /recipes/[id].vue
- Update route references in browse and search pages
- Preserve full functionality of recipe details page
- Maintain existing component structure and interactions
- Improve routing consistency across the application
