+++ pages/browse.vue
@@ -197,7 +197,7 @@
               v-for="recipe in filteredRecipes" 
               :key="recipe.id"
               class="p-4 rounded-2xl transition-all duration-200 bg-surface hover:bg-surface-hover border-theme cursor-pointer"
-              @click="router.push(`/${recipe.id}`)"
+              @click="router.push(`/recipes/${recipe.id}`)"
             >
               <div class="flex items-start gap-4">
                 <!-- Recipe Image -->
